requirements.txt
pytest
requests
requests-mock
jwt
cachetools
python-dotenv